+++ changelog.md
@@ -5,7 +5,7 @@
 - jcontent_by_categories : caché l'article en cours et article en cours indéterminé : essayer de le récupérer par getInput
 - sql : ajout pagination, setlimit accepte un nombre ou une requête sql
 - sql : ajout de route-format : calcul du lien vers un article
-- upbtn_makefile : ajout de 
+- upbtn_makefile : ajout de page-load-status en cas d'attente de chargement d'une action
 
 #### Modification interne