+++ changelog.md
@@ -1,4 +1,4 @@
-## 10/11/2025 - version 5.4.5
+## 22/11/2025 - version 5.4.5
 
 #### Modifications actions